feature/backend: Se resuelven conflictos Hibernate, se añade un entrypint.sh para asegurar la conexion con la db y se valida con la ruta de documentacion swagger.

+++ Desarrollo/backend/src/main/java/com/cuidadoeterno/backend/modules/finanzas/model/CuentaBanco.java
@@ -21,7 +21,7 @@ public class CuentaBanco {
     @Id
     @GeneratedValue(strategy = GenerationType.IDENTITY)
     @Column(name = "id_cuenta")
-    private Long idCuenta;
+    private Integer idCuenta;
 
     /**
      * FK → CUIDADOR.id_persona
